Combine @is and @provides
These two directives are functionally doing the same thing we just limit them to different use cases with the directive locations.

They both tell the query planner that if you go through this path some of the data you will get back can be sourced either from an argument or it will be provided by the subgraph

I am open for discussion of not merging them. But at second glance they do seem very similar. It also helps with the short and unclear name of `@is`